+++ crates/medea-reactive/src/subscribers_store/progressable/mod.rs
@@ -1,26 +1,17 @@
 //! Progressable [`SubscribersStore`].
 
 pub mod guarded;
-pub mod recheckable_future;
+pub mod processed;
 
-use std::{
-    cell::RefCell,
-    fmt,
-    pin::Pin,
-    rc::Rc,
-    task::{Context, Poll},
-};
+use std::{cell::RefCell, rc::Rc};
 
-use futures::{
-    channel::mpsc, future::LocalBoxFuture, stream::LocalBoxStream, Future,
-    FutureExt,
-};
+use futures::{channel::mpsc, stream::LocalBoxStream};
 
 use crate::{subscribers_store::SubscribersStore, ObservableCell};
 
 pub use self::{
     guarded::{Guard, Guarded},
-    recheckable_future::RecheckableFutureExt,
+    processed::{AllProcessed, Processed},
 };
 
 /// [`SubscribersStore`] for progressable collections/field.
@@ -55,8 +46,14 @@ impl<T> SubStore<T> {
     /// Returns [`Future`] resolving when all subscribers processes update.
     ///
     /// [`Future`]: std::future::Future
-    pub fn when_all_processed(&self) -> RecheckableCounterFuture {
-        RecheckableCounterFuture::new(Rc::clone(&self.counter))
+    pub fn when_all_processed(&self) -> Processed<'static, ()> {
+        let counter = Rc::clone(&self.counter);
+        Processed::new(Box::new(move || {
+            let counter = Rc::clone(&counter);
+            Box::pin(async move {
+                let _ = counter.when_eq(0).await;
+            })
+        }))
     }
 }
 
@@ -79,64 +76,5 @@ where
     #[inline]
     fn wrap(&self, value: T) -> Guarded<T> {
         Guarded::wrap(value, Rc::clone(&self.counter))
-    }
-}
-
-/// [`RecheckableFutureExt`] for [`SubStore::subscribe`].
-pub struct RecheckableCounterFuture {
-    /// Reference to the [`SubStore::counter`].
-    counter: Rc<ObservableCell<u32>>,
-
-    /// Current [`Future`] which will be polled on
-    /// [`RecheckableCounterFuture::poll`].
-    pending_fut: Option<LocalBoxFuture<'static, ()>>,
-}
-
-impl RecheckableFutureExt for RecheckableCounterFuture {
-    /// Returns `true` if [`RecheckableCounterFuture::counter`] is `0`.
-    fn is_done(&self) -> bool {
-        self.counter.get() == 0
-    }
-
-    /// Refreshes [`RecheckableCounterFuture::pending_fut`] with a new
-    /// [`RecheckableCounterFuture::counter`]'s [`ObservableCell::when_eq`]
-    /// [`Future`].
-    fn restart(&mut self) {
-        self.pending_fut = Some(Box::pin(self.counter.when_eq(0).map(|_| ())));
-    }
-}
-
-impl fmt::Debug for RecheckableCounterFuture {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        f.debug_struct("RecheckableCounterFuture")
-            .field("counter", &self.counter)
-            .finish()
-    }
-}
-
-impl RecheckableCounterFuture {
-    /// Returns new [`RecheckableCounterFuture`] for the provided counter.
-    pub(super) fn new(counter: Rc<ObservableCell<u32>>) -> Self {
-        Self {
-            pending_fut: None,
-            counter,
-        }
-    }
-}
-
-impl Future for RecheckableCounterFuture {
-    type Output = ();
-
-    #[allow(clippy::option_if_let_else)]
-    fn poll(
-        mut self: Pin<&mut Self>,
-        cx: &mut Context<'_>,
-    ) -> Poll<Self::Output> {
-        if let Some(fut) = self.pending_fut.as_mut() {
-            fut.as_mut().poll(cx)
-        } else {
-            self.restart();
-            self.poll(cx)
-        }
     }
 }